+++ src/main/java/kitchenpos/dto/OrderLineItemDto.java
@@ -1,6 +1,6 @@
-package kitchenpos.application.dto;
+package kitchenpos.dto;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemDto {
