+++ src/main/java/kitchenpos/order/repository/OrderLineItemRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.order.repository;
 
 import java.util.List;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
